Added decode_ic(), the InputCode to Keys converter.
